Finish plot_abins.py

+++ plot_abins.py
@@ -88,6 +88,9 @@ def has_data_started(line):
 
 
 def normalise_data(abins_x, abins_y, experimental):
+    #abins_y -= abins_y[-1]
+    #experimental[:, 1] -= experimental[-1, 1]
+    
     abins_start = np.where(abins_x > 50)[0][0]
     exp_start = np.where(experimental[:, 0] > 50)[0][0]
 
@@ -174,20 +177,26 @@ if __name__ == '__main__':
 
         print(compound)
 
-        result = Abins(VibrationalOrPhononFile=os.path.join(directory, f'{compound}-phonopy.yaml'),
-                       AbInitioProgram='FORCECONSTANTS',
-                       Instrument='TOSCA',
-                       SumContributions=True,
-                       QuantumOrderEventsNumber='2',
-                       Autoconvolution=True,
-                       Setting='All detectors (TOSCA)')
+        if os.path.exists(os.path.join(directory, f'abins.npy')):
+            result = np.load(os.path.join(directory, 'abins.npy'))
+            energy = result[0, :]
+            s = result[1, :]
+        else:
+            result = Abins(VibrationalOrPhononFile=os.path.join(directory, f'{compound}-phonopy.yaml'),
+                           AbInitioProgram='FORCECONSTANTS',
+                           Instrument='TOSCA',
+                           SumContributions=True,
+                           QuantumOrderEventsNumber='2',
+                           Autoconvolution=True,
+                           Setting='All detectors (TOSCA)',
+                           ScaleByCrossSection="Total")
 
-        energy = result[0].extractX().flatten()
-        energy = (energy[1:] + energy[:-1]) / 2
-        s = result[0].extractY().flatten()
+            energy = result[0].extractX().flatten()
+            energy = (energy[1:] + energy[:-1]) / 2
+            s = result[0].extractY().flatten()
 
-        np.save(os.path.join(directory, 'abins.npy'),
-                np.stack([energy, s]))
+            np.save(os.path.join(directory, 'abins.npy'),
+                    np.stack([energy, s]))
 
         ins_data = parse_data_file(os.path.join(INS_DIR, f'{compound}.dat'))
         try:
@@ -197,27 +206,35 @@ if __name__ == '__main__':
                 print(val)
             raise
 
-        energy, s, ins_data, maximum = normalise_data(energy, s, ins_data)
+        energy, s, ins_data, y_max = normalise_data(energy, s, ins_data)
 
-        fig, ax = plt.subplots(dpi=400)
+        fig, ax = plt.subplots(dpi=2000)
 
-        ax.plot(ins_data[:, 0], ins_data[:, 1], label='Experimental', alpha=0.6)
-        ax.plot(energy, s, label='AbINS', alpha=0.6)
+        ax.plot(ins_data[:, 0], ins_data[:, 1], label='Experimental', alpha=0.7, c='#1E5DF8', linewidth=2.5)
+        ax.plot(energy, s, label='AbINS', alpha=0.7, c='#E94D36', linewidth=2.5)
 
-        ax.set_xlabel('Energy transfer $(cm^{-1})$')
-        ax.set_ylabel('S(q, w)')
+        ax.set_xlabel('Energy transfer $(cm^{-1})$', fontsize=20)
+        ax.set_ylabel('S(q, w)', fontsize=20)
 
         ax.set_xlim(0, 4000)
-        ax.set_ylim(top=maximum*1.5)
+        
+        if np.max(ins_data[:, 1]) > 3 * y_max:
+            y_min = min([np.min(s), np.min(ins_data[:, 1])])
+            ax.set_ylim(y_min*0.9, y_max*1.5)
 
+        ax.tick_params(length=5, width=2, labelsize=15)
         ax.axes.get_yaxis().set_ticks([])
 
-        plt.legend()
+        plt.legend(fontsize=15)
 
+        fig.tight_layout()
         fig.savefig(os.path.join(directory, f'{compound}.png'))
         plt.close(fig)
 
-        result.delete()
+        try:
+            result.delete()
+        except (NameError, AttributeError):
+            pass
 
     created_hdf_files = glob.glob(os.path.join(HOME_DIR, '*.hdf5'))
     for file in created_hdf_files:
